+++ src/pages/workspaces/workspace/analysis/Analyses.test.ts
@@ -17,12 +17,12 @@ import {
   AnalysesProps,
   BaseAnalyses, getUniqueFileName
 } from 'src/pages/workspaces/workspace/analysis/Analyses'
+import { analysisLauncherTabName } from 'src/pages/workspaces/workspace/analysis/runtime-common-components'
+import { AnalysisFile, getFileFromPath, useAnalysisFiles } from 'src/pages/workspaces/workspace/analysis/useAnalysisFiles'
 import {
   AbsolutePath, FileName, findPotentialNotebookLockers, notebookLockHash
-} from 'src/pages/workspaces/workspace/analysis/file-utils'
-import { analysisLauncherTabName } from 'src/pages/workspaces/workspace/analysis/runtime-common-components'
-import { toolLabels } from 'src/pages/workspaces/workspace/analysis/tool-utils'
-import { AnalysisFile, getFileFromPath, useAnalysisFiles } from 'src/pages/workspaces/workspace/analysis/useAnalysisFiles'
+} from 'src/pages/workspaces/workspace/analysis/utils/file-utils'
+import { toolLabels } from 'src/pages/workspaces/workspace/analysis/utils/tool-utils'
 import { asMockedFn } from 'src/testing/test-utils'
 
 
@@ -33,9 +33,9 @@ jest.mock('src/libs/nav', (): NavExports => ({
   goToPath: jest.fn(),
 }))
 
-type FileUtilsExports = typeof import('src/pages/workspaces/workspace/analysis/file-utils')
-jest.mock('src/pages/workspaces/workspace/analysis/file-utils', (): FileUtilsExports => ({
-  ...jest.requireActual('src/pages/workspaces/workspace/analysis/file-utils'),
+type FileUtilsExports = typeof import('src/pages/workspaces/workspace/analysis/utils/file-utils')
+jest.mock('src/pages/workspaces/workspace/analysis/utils/file-utils', (): FileUtilsExports => ({
+  ...jest.requireActual('src/pages/workspaces/workspace/analysis/utils/file-utils'),
   notebookLockHash: jest.fn(),
   findPotentialNotebookLockers: jest.fn()
 }))
