+++ src/pages/workspaces/workspace/analysis/modals/AnalysisModal.ts
@@ -21,21 +21,21 @@ import { usePrevious, withDisplayName } from 'src/libs/react-utils'
 import * as Style from 'src/libs/style'
 import * as Utils from 'src/libs/utils'
 import { BaseWorkspace, cloudProviderTypes, isGoogleWorkspaceInfo } from 'src/libs/workspace-utils'
-import { getCurrentApp } from 'src/pages/workspaces/workspace/analysis/app-utils'
-import { getCurrentPersistentDisk } from 'src/pages/workspaces/workspace/analysis/disk-utils'
-import { getFileName } from 'src/pages/workspaces/workspace/analysis/file-utils'
 import { AzureComputeModalBase } from 'src/pages/workspaces/workspace/analysis/modals/AzureComputeModal'
 import { ComputeModalBase } from 'src/pages/workspaces/workspace/analysis/modals/ComputeModal'
 import { CromwellModalBase } from 'src/pages/workspaces/workspace/analysis/modals/CromwellModal'
 import { GalaxyModalBase } from 'src/pages/workspaces/workspace/analysis/modals/GalaxyModal'
 import { WarningTitle } from 'src/pages/workspaces/workspace/analysis/modals/WarningTitle'
-import { analysisNameInput, analysisNameValidator, baseRmd, notebookData } from 'src/pages/workspaces/workspace/analysis/notebook-utils'
+import { AnalysisFileStore } from 'src/pages/workspaces/workspace/analysis/useAnalysisFiles'
+import { getCurrentApp } from 'src/pages/workspaces/workspace/analysis/utils/app-utils'
+import { getCurrentPersistentDisk } from 'src/pages/workspaces/workspace/analysis/utils/disk-utils'
+import { getFileName } from 'src/pages/workspaces/workspace/analysis/utils/file-utils'
+import { analysisNameInput, analysisNameValidator, baseRmd, notebookData } from 'src/pages/workspaces/workspace/analysis/utils/notebook-utils'
+import { isResourceDeletable } from 'src/pages/workspaces/workspace/analysis/utils/resource-utils'
 import {
   getCurrentRuntime
-} from 'src/pages/workspaces/workspace/analysis/runtime-utils'
-import { AppTool, cloudAppTools, cloudRuntimeTools, getAppType, getToolLabelFromFileExtension, getToolLabelFromRuntime, isAppToolLabel, runtimeTools, Tool, toolExtensionDisplay, toolLabels, tools } from 'src/pages/workspaces/workspace/analysis/tool-utils'
-import { AnalysisFileStore } from 'src/pages/workspaces/workspace/analysis/useAnalysisFiles'
-import { isResourceDeletable } from 'src/pages/workspaces/workspace/analysis/utils/resource-utils'
+} from 'src/pages/workspaces/workspace/analysis/utils/runtime-utils'
+import { AppTool, cloudAppTools, cloudRuntimeTools, getAppType, getToolLabelFromFileExtension, getToolLabelFromRuntime, isAppToolLabel, runtimeTools, Tool, toolExtensionDisplay, toolLabels, tools } from 'src/pages/workspaces/workspace/analysis/utils/tool-utils'
 import validate from 'validate.js'
