+++ src/pages/workspaces/workspace/analysis/modals/AzureComputeModal.test.js
@@ -9,8 +9,8 @@ import { formatUSD } from 'src/libs/utils'
 import {
   defaultAzureWorkspace, imageDocs, testAzureDefaultRegion
 } from 'src/pages/workspaces/workspace/analysis/_testData/testData'
-import { getAzureComputeCostEstimate, getAzureDiskCostEstimate } from 'src/pages/workspaces/workspace/analysis/cost-utils'
-import { toolLabels } from 'src/pages/workspaces/workspace/analysis/tool-utils'
+import { getAzureComputeCostEstimate, getAzureDiskCostEstimate } from 'src/pages/workspaces/workspace/analysis/utils/cost-utils'
+import { toolLabels } from 'src/pages/workspaces/workspace/analysis/utils/tool-utils'
 import { asMockedFn } from 'src/testing/test-utils'
 
 import { AzureComputeModalBase } from './AzureComputeModal'
