+++ src/pages/workspaces/workspace/analysis/modals/ExportAnalysisModal.js
@@ -13,9 +13,9 @@ import { FormLabel } from 'src/libs/forms'
 import * as Nav from 'src/libs/nav'
 import { useCancellation } from 'src/libs/react-utils'
 import * as Utils from 'src/libs/utils'
-import { getAnalysisFileExtension, getDisplayName, getExtension, stripExtension } from 'src/pages/workspaces/workspace/analysis/file-utils'
-import { analysisNameInput, analysisNameValidator } from 'src/pages/workspaces/workspace/analysis/notebook-utils'
 import { analysisLauncherTabName, analysisTabName } from 'src/pages/workspaces/workspace/analysis/runtime-common-components'
+import { getAnalysisFileExtension, getDisplayName, getExtension, stripExtension } from 'src/pages/workspaces/workspace/analysis/utils/file-utils'
+import { analysisNameInput, analysisNameValidator } from 'src/pages/workspaces/workspace/analysis/utils/notebook-utils'
 import validate from 'validate.js'
