+++ src/pages/workspaces/workspace/useWorkspace.test.ts
@@ -9,7 +9,7 @@ import { getConfig } from 'src/libs/config'
 import * as Notifications from 'src/libs/notifications'
 import { getUser, workspaceStore } from 'src/libs/state'
 import { DeepPartial } from 'src/libs/type-utils/deep-partial'
-import { defaultLocation } from 'src/pages/workspaces/workspace/analysis/runtime-utils'
+import { defaultLocation } from 'src/pages/workspaces/workspace/analysis/utils/runtime-utils'
 import {
   azureBucketRecheckRate, googlePermissionsRecheckRate, useWorkspace
 } from 'src/pages/workspaces/workspace/useWorkspace'
